backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.16
langgraph==0.0.26
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
reportlab==4.4.10
structlog==23.2.0
psutil==5.9.6
numpy==1.26.4
requests==2.32.5
